Update Bitski.js to load dynamically
Updates Bitski.js to load the SDK dynamically, so we can make updates
more quickly without requiring consumers to bump the package as well.
The implementation is still in the same package so we can expose the
types to consumers, but the SDK has been moved to the `/-private`
directory which signals that no one should import it directly.